Ensure upgrade keyring exists prior to upgrade checks
During ceph to ceph-osd/ceph-mon migrations, the bootstrap keyring
for the cluster will be in place as the ceph-osd units are started
alongside existing ceph units.

Switch this check to look for the upgrade keyring, which won't be
in place until the ceph-osd <-> ceph-mon relation is complete, at
which point in time a) the unit has the correct access to perform
the upgrade and b) the previous/current version check code will
not trip over due to the previous value of the source option
being None, resulting in a fallback to 'distro' as the previous
source of ceph.

Change-Id: I10895c60aeb543a10461676e4455ed6b5e2fdb46
Closes-Bug: 1729369