Change `osd crush location` to `crush location`
Upstream Ceph removed the `osd crush location` option in
commit f9db479a14d9103a2b7c0a24d958fe5fff94100e [0]. This
causes new clusters deployed from the Pike UCA (Luminous)
using the customize-failure-domain option to not create or
move the OSD to the correct spot in the OSD tree. The end
result is that placement groups will fail to peer because
there are no racks to select hosts and OSDs from.

Instead, the charm should set the more generic `crush
location` option in the ceph.conf file. It is and has been
supported since the Firefly (trusty-icehouse) version of
Ceph.

[0] https://github.com/ceph/ceph/commit/f9db479a14d9103a2b7c0a24d958fe5fff94100e

Change-Id: I0b7055b20f54096a2f33583079326aee17726355
Closes-Bug: 1730839