Update tox.ini to stop using unverified package
As of pip 10.0, --allow-unverified is not permitted.

Use of the flag in this repo was previously used to force
installation of python-apt to accommodate certain unit tests.

The unverified package, python-apt, is no longer necessary
for test execution.

Related-Bug: #1760720

Change-Id: I06e5afd0bfb627b2de64e93e51c182cf881e38d7
(cherry picked from commit e179d3680227728307dce7bb05dbbf0a8ab70274)