Remove replace-osd action
At present this action does not work.  While looking to repair the
functionality I found a number of issues with the current
implementation.

For now I suggest we remove this functionality, and at some point we may
consider replacing it with a `remove-disk` action.

Sync in relevant changes from charms.ceph

Depends-On: Id61b87927c43d807aacc93cf05ec8f88d91b7a39
Change-Id: Ic71d304ff65a05ab7249f4dd07adc45429a323e9