add-disk: Ensure key-manager config is passed to osdize
Recent changes to support vault for key management require that
the 'osd-encrypt-keymanager' is passed to all osdize calls so
that the correct key management approach is taken.

Ensure that the add-disk action does the same, otherwise keys
will always be stored in the ceph mon KV store, rather than
in Vault.

Closes-Bug: 1780920

Change-Id: I8c722d38d68f13dc00c7444a50d67ce37fbd6a29