Notify MON cluster of number of bootstrapped OSD's
To allow the ceph-mon charm to better assess when the Ceph cluster
is in a usable state, provide the number of OSD devices that where
bootstrapped into the Ceph cluster over the relation to ceph-mon.

This is used by the ceph-mon charm inconjunction with the
'expected-osd-count' configuration option to delay pool creation
and issue of keys for clients until the expected number of OSD's
have been bootstrapped into the cluster.

Change-Id: I1370524f0f31120e3cb7305c5bc509a6494c5586
Closes-Bug: 1794878
(cherry picked from commit 63f9ac2c7cd0db8f212bba6d278af2b0316b7760)