Update cloud-archive.list when upgrading from Pike to Queens
ceph-osd charm only upgrades when the ceph version changes, for the
case of upgrading from Pike to Queens the charm is skipping any
upgrades, because the Cloud Archive has Luminous for those 2 releases.

This patch checks if the requested ceph version is luminous and if the
'source' changed from pike to queens to then upgrade
/etc/apt/sources.list.d/cloud-archive.list via add_source()

As part of this backport, I manually un-picked the stein dict
entry from lib/ceph/utils.py, as it is not currently applicable to
the stable charm.

Change-Id: I27c2d7648a1add6528924724a03682060d2d6007
Closes-Bug: 1778823
(cherry picked from commit 336581efeb9a1c5d363c5d3ce631d9bad1a8a337)