bluestore-wal only needed if separate to DB device
Update config.yaml to clarify that bluestore-wal should only be set
where a separate (faster) device is being used for the WAL. Otherwise,
the WAL is automatically maintained within the space of the DB device
and does not need to be configured separately.

Additionally clarify that this device is used as an LVM PV and space is
allocated for each block device based on the
bluestore-block-{db,wal}-size setting.

Change-Id: I54fc582ecb2cee5de1302685e9103c636c7a307b
Ref: http://docs.ceph.com/docs/master/rados/configuration/bluestore-config-ref/