Re-fix Traceback issue when ceph-osd upgrade fails
This was originally fixed in Ib7d1fdc8f91bc992ccf618ef6f57e99bb90c2dbc
but unfortunately wasn't also added to the charms.ceph library.  Thus,
this is a re-application of that fix; the charms to ceph fix is in [1].

Bug/1770740 surfaced an issue where get_upgrade_position() returns None
but the calling function expects and exception to the thrown if the
"None" condition exists.  This just fixes the code so that the Traceback
is stopped and the appropriate error/message is logged for the
condition.

[1] https://review.opendev.org/#/c/695163/
    I16539b2bc35104eed54033bebb1154cad8a5cf0f

Change-Id: Ieee8d13f25027ad540a23a6428c2226b6c20999a
Partial-Bug: #1770740