Ensure python3-hvac is installed for charms with encypt option
The referenced bug is essentially: make vault:secrets relation to vault
but keep the 'encrypt' option as False. In this case, the Context
handling code in charm-helpers is expecting python3-hvac to be
available, but it is only installed if the encrypt option is set to
True.  Hence the charm crashes.  This resolves that crash.

Note the related charm-helpers fix [1].

[1]: https://github.com/juju/charm-helpers/pull/431

Change-Id: I9cb60a9340554c91668272b46f7c2dcf9f0ac2d1
Closes-bug: #1862085