Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I0fb5e14bd2e58d4abdcce6e230d5c4c1a1d00af7
(cherry picked from commit af700cf3261c99e826070f19bf70da3b516ee99c)