Implement key rotation for OSD
This patchset implements key rotation for OSD's. It does so by
receiving an update in the relation data bag from ceph-mons (where
the actions is started), which informs the OSD units for which
OSD the key needs to be rotated and the key itself. The OSD units
then check if they are managing the ID specified, and if so,
proceed to rotate the key.

Change-Id: I382a0a657b31c172a036ce7ca62facbbce32b4a0