Add antelope to the UCA codename map
According to https://docs.openstack.org/charm-guide/latest/project/charm-delivery.html#tracks-for-the-openstack-charms-project
quincy is the supported release for OpenStack 2023.1 (antelope),
so antelope must be included in the UCA codename map.

Otherwise the charm cannot determine correctly what action
should be taken after upgrading the `source` to antelope.
This can result in the charm potentially attempting to upgrade
to version `None` and crashing.

Change-Id: Ic9f6c5260e1396980b7a5316484ca5e970172de9